fix: abstracted furnace logic to an abstract
